Dependency-inject the driver into the registry mock. This allows me to require the registry and supply a different driver. Possibly a first step to factor out the registry code into another module for re-uese.
